Time: 179 ms (5.14%), Space: 18.8 MB (74.90%) - LeetHub

+++ 875-koko-eating-bananas/875-koko-eating-bananas.cpp
@@ -9,7 +9,7 @@ public:
         int high=piles[n-1]+1;
         int l=1;
     
-    int count=0;
+    long long count=0;
         while(l<=high)
         {
             int mini=(high+l)/2;
@@ -17,19 +17,15 @@ public:
             for(int i=0;i<n;i++)
             {
                 count+=ceil(1.0*piles[i]/mini);
-                // cout<<count<<endl;
             }
             if(count>h)
             { 
-                // j+=1;
                 l=mini+1;
             }
-           // else if(count<h)
-           //      high=mini-1;
+        
             else
                 high=mini-1;
-                // return mini;
-                // return mini;
+         
             
         }
         return l;
